Do initial adjustments

+++ src/main/java/com/finalproject/order_service/Repo/WishlistRepository.java
@@ -7,6 +7,6 @@ import java.util.List;
 
 public interface WishlistRepository extends JpaRepository<Wishlist,Long> {
     List<Wishlist> findByUserId(Long userId);
-    Wishlist findByUserIdAndProductId(Long userId, Long productId);
+    Wishlist findByUserIdAndProductId(Long userId, Integer productId);
 
 }
